DNGO-fastapi/requirements.txt
# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
pg8000
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# CORS
python-dotenv==1.0.0

# Utilities
httpx==0.26.0
requests==2.31.0
apscheduler==3.11.2
firebase-admin==6.5.0